moved extract zip fils functions into workers
